Moved target read/write nbt code to target helper.
